Database/07_REST API/DRF/requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.4
debugpy==1.6.2
decorator==5.1.0
Django==3.2.13
django-appconf==1.0.5
django-bootstrap-v5==1.0.2
django-debug-toolbar==3.2.4
django-extensions==3.2.1
django-imagekit==4.1.0
django-seed==0.2.2
djangorestframework==3.14.0
entrypoints==0.4
executing==0.8.3
Faker==13.3.5
idna==3.4
importlib-metadata==2.1.1
ipykernel==6.15.1
ipython==7.27.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==7.3.4
jupyter-core==4.11.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==9.2.0
prompt-toolkit==3.0.20
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.10.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
pywin32==304
pyzmq==23.2.0
requests==2.28.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.3
stack-data==0.2.0
text-unidecode==1.3
tornado==6.2
traitlets==5.1.0
tzdata==2021.5
urllib3==1.26.12
wcwidth==0.2.5
zipp==3.4.1
